vcs-to-changelog: Allow loading of custom quirks file
gnulib does not have a quirks file and if the scripts are not copied
over to the project tree, it may never find the quirks file in the
right place.  Add a flag to vcs_to_changelog.py to allow one to
specify the location of the quirks file instead.

	* build-aux/vcs_to_changelog.py: New commandline option -q.