obstack: use presence of uintptr_t
The code depends on the presence of uintptr_t,
not on whether ptrdiff_t is narrower than a pointer,
so use C macros that reflect this.
Also, clean up the namespace in the typical GCC-like case
* lib/obstack.in.h (_OBSTACK_UINTPTR_TYPE): New macro.
Include <stdint.h> only if needed (i.e., not GCC-like).
(__PTR_ALIGN): Use new macro instead of relying on
__GL_SMALL_PTRDIFF_T__ and directly on uintptr_t.
* m4/obstack.m4 (gl_FUNC_OBSTACK): Simplify by removing
__GL_SMALL_PTRDIFF_T__.